+++ examples/js/selection/custom-multi-select-table.js
@@ -1,0 +1,90 @@
+/* eslint max-len: 0 */
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { BootstrapTable, TableHeaderColumn } from 'react-bootstrap-table';
+require('../../customMultiSelect.css');
+
+const products = [];
+
+function addProducts(quantity) {
+  const startId = products.length;
+  for (let i = 0; i < quantity; i++) {
+    const id = startId + i;
+    products.push({
+      id: id,
+      name: 'Item name ' + id,
+      price: 2100 + i
+    });
+  }
+}
+
+addProducts(5);
+
+class Checkbox extends React.Component {
+  componentDidMount() { this.update(this.props.checked); }
+  componentWillReceiveProps(props) { this.update(props.checked); }
+  update(checked) {
+    ReactDOM.findDOMNode(this).indeterminate = checked === 'indeterminate';
+  }
+  render() {
+    return (
+      <input className='react-bs-select-all'
+        type='checkbox'
+        name={ 'checkbox' + this.props.rowIndex }
+        id={ 'checkbox' + this.props.rowIndex }
+        checked={ this.props.checked }
+        onChange={ this.props.onChange } />
+    );
+  }
+}
+
+export default class CustomMultiSelectTable extends React.Component {
+  customMultiSelect(props) {
+    const { type, checked, disabled, onChange, rowIndex } = props;
+    /*
+    * If rowIndex is 'Header', means this rendering is for header selection column.
+    */
+    if (rowIndex === 'Header') {
+      return (
+        <div className='checkbox-personalized'>
+          <Checkbox {...props}/>
+          <label htmlFor={ 'checkbox' + rowIndex }>
+            <div className='check'></div>
+          </label>
+        </div>);
+    } else {
+      return (
+        <div className='checkbox-personalized'>
+          <input
+            type={ type }
+            name={ 'checkbox' + rowIndex }
+            id={ 'checkbox' + rowIndex }
+            checked={ checked }
+            disabled={ disabled }
+            onChange={ e=> onChange(e, rowIndex) }
+            ref={ input => {
+              if (input) {
+                input.indeterminate = props.indeterminate;
+              }
+            } }/>
+          <label htmlFor={ 'checkbox' + rowIndex }>
+            <div className='check'></div>
+          </label>
+        </div>);
+    }
+  }
+  render() {
+    const selectRowProp = {
+      mode: 'checkbox',
+      customComponent: this.customMultiSelect
+    };
+    return (
+      <BootstrapTable data={ products } selectRow={ selectRowProp }
+        tableHeaderClass='custom-select-header-class' tableBodyClass='custom-select-body-class'>
+          <TableHeaderColumn dataField='id' isKey={ true }>Product ID</TableHeaderColumn>
+          <TableHeaderColumn dataField='name'>Product Name</TableHeaderColumn>
+          <TableHeaderColumn dataField='price'>Product Price</TableHeaderColumn>
+      </BootstrapTable>
+    );
+  }
+}