+++ examples/js/selection/demo.js
@@ -12,6 +12,7 @@ import SelectValidationTable from './select-validation-table';
 import RowClickTable from './row-click-table';
 import OnlySelectedTable from './only-show-selected-table';
 import ExternallyManagedSelection from './externally-managed-selection';
+import CustomMultiSelectTable from './custom-multi-select-table';
 import SelectAll from './all-select';
 
 class Demo extends React.Component {
@@ -33,6 +34,15 @@ class Demo extends React.Component {
             <div className='panel-body'>
               <h5>Source in /examples/js/selection/multi-select-table.js</h5>
               <MultiSelectTable />
+            </div>
+          </div>
+        </div>
+        <div className='col-md-offset-1 col-md-8'>
+          <div className='panel panel-default'>
+            <div className='panel-heading'>Custom Multi Select Example</div>
+            <div className='panel-body'>
+              <h5>Source in /examples/js/selection/custom-multi-select-table.js</h5>
+              <CustomMultiSelectTable />
             </div>
           </div>
         </div>