+++ src/BootstrapTable.js
@@ -268,6 +268,7 @@ class BootstrapTable extends Component {
             tableHeaderClass={ this.props.tableHeaderClass }
             style={ this.props.headerStyle }
             rowSelectType={ this.props.selectRow.mode }
+            customComponent={ this.props.selectRow.customComponent }
             hideSelectColumn={ this.props.selectRow.hideSelectColumn }
             sortName={ sortInfo ? sortInfo.sortField : undefined }
             sortOrder={ sortInfo ? sortInfo.order : undefined }
@@ -478,6 +479,7 @@ class BootstrapTable extends Component {
   }
 
   handleEditCell(newVal, rowIndex, colIndex) {
+    const { onCellEdit } = this.props.options;
     const { beforeSaveCell, afterSaveCell } = this.props.cellEdit;
     let fieldName;
     React.Children.forEach(this.props.children, function(column, i) {
@@ -495,6 +497,17 @@ class BootstrapTable extends Component {
         });
         return;
       }
+    }
+
+    if (onCellEdit) {
+      onCellEdit(this.state.data[rowIndex], fieldName, newVal);
+    }
+
+    if (this.isRemoteDataSource()) {
+      if (afterSaveCell) {
+        afterSaveCell(this.state.data[rowIndex], fieldName, newVal);
+      }
+      return;
     }
 
     const result = this.store.edit(newVal, rowIndex, fieldName).get();
@@ -881,6 +894,13 @@ class BootstrapTable extends Component {
         header.childNodes[i].style.width = result;
         header.childNodes[i].style.minWidth = result;
       }
+    } else {
+      React.Children.forEach(this.props.children, (child, i) => {
+        if (child.props.width) {
+          header.childNodes[i].style.width = `${child.props.width}px`;
+          header.childNodes[i].style.minWidth = `${child.props.width}px`;
+        }
+      });
     }
   }
 
@@ -934,6 +954,7 @@ BootstrapTable.propTypes = {
       Const.ROW_SELECT_SINGLE,
       Const.ROW_SELECT_MULTI
     ]),
+    customComponent: PropTypes.func,
     bgColor: PropTypes.string,
     selected: PropTypes.array,
     onSelect: PropTypes.func,
@@ -988,6 +1009,7 @@ BootstrapTable.propTypes = {
     onSearchChange: React.PropTypes.func,
     onAddRow: React.PropTypes.func,
     onExportToCSV: React.PropTypes.func,
+    onCellEdit: React.PropTypes.func,
     noDataText: PropTypes.oneOfType([ PropTypes.string, PropTypes.object ]),
     handleConfirmDeleteRow: PropTypes.func,
     prePage: PropTypes.string,
@@ -1034,7 +1056,8 @@ BootstrapTable.defaultProps = {
     hideSelectColumn: false,
     clickToSelectAndEditCell: false,
     showOnlySelected: false,
-    unselectable: []
+    unselectable: [],
+    customComponent: undefined
   },
   cellEdit: {
     mode: Const.CELL_EDIT_NONE,